requirements.txt
requests
certifi
truststore
customtkinter
pillow
pandas
wikipedia
google-generativeai
pyworms
supabase
pydantic
python-dotenv
lpsn>=1.0.0
beautifulsoup4
